refactor: html 파일 패키지 이름 변경
