+++ src/main/java/com/example/jomajomadelivery/account/auth/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.example.jomajomadelivery.auth.controller;
+package com.example.jomajomadelivery.account.auth.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
